Блок 2 выполнен.

+++ src/main/java/com/github/spind30/starbankapp/services/RuleService.java
@@ -1,47 +1,59 @@
 package com.github.spind30.starbankapp.services;
 
+import com.github.spind30.starbankapp.dto.DynamicRuleDTO;
 import com.github.spind30.starbankapp.model.rule.DynamicRule;
 import com.github.spind30.starbankapp.repository.RuleRepository;
 import jakarta.transaction.Transactional;
-import lombok.AllArgsConstructor;
+import lombok.RequiredArgsConstructor;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
 import java.util.UUID;
+import java.util.stream.Collectors;
 
 @Service
-@AllArgsConstructor
+@RequiredArgsConstructor
 public class RuleService {
 
-    private final Logger logger = LoggerFactory.getLogger(RuleService.class);
+    private static final Logger logger = LoggerFactory.getLogger(RuleService.class);
 
-    @Autowired
     private final RuleRepository ruleRepository;
 
     @Transactional
-    public DynamicRule createRule(DynamicRule dynamicRule) {
+    public DynamicRuleDTO createRule(DynamicRuleDTO dynamicRuleDTO) {
+        DynamicRule dynamicRule = dynamicRuleDTO.toEntity();
         if (dynamicRule.getRule() != null) {
             dynamicRule.getRule().forEach(query -> query.setDynamicRule(dynamicRule));
         }
-        logger.info("Правило сохранено");
-        return ruleRepository.save(dynamicRule);
+        DynamicRule savedRule = ruleRepository.save(dynamicRule);
+        logger.info("Правило сохранено: {}", savedRule.getProductId());
+        return DynamicRuleDTO.fromEntity(savedRule);
     }
-
 
     public void deleteRule(UUID id) {
-        ruleRepository.deleteById(id);
-    }
-
-    public DynamicRule editRule(DynamicRule dynamicRule) {
-       return ruleRepository.save(dynamicRule);
-    }
-
-    public List<DynamicRule> getAllRules() {
-        return ruleRepository.findAll();
+        if (ruleRepository.existsById(id)) {
+            ruleRepository.deleteById(id);
+            logger.info("Правило удалено: {}", id);
+        } else {
+            logger.warn("Правило с ID {} не найдено для удаления.", id);
+        }
     }
 
 
+    @Transactional
+    public DynamicRuleDTO editRule(DynamicRuleDTO dynamicRuleDTO) {
+        DynamicRule updatedRule = ruleRepository.save(dynamicRuleDTO.toEntity());
+        logger.info("Правило обновлено: {}", updatedRule.getProductId());
+        return DynamicRuleDTO.fromEntity(updatedRule);
+    }
+
+    public List<DynamicRuleDTO> getAllRules() {
+        return ruleRepository.findAll()
+                .stream()
+                .map(DynamicRuleDTO::fromEntity)
+                .collect(Collectors.toList());
+    }
 }
+
